# [ISSUE-3] Cache by entity and constraint definitions