--- a/README.md
+++ b/README.md
@@ -48,7 +48,6 @@ With code coverage:
 make test-coverage
 ```
 
-
 ### Linting
 
 ```bash