--- a/requirements/requirements-dev.txt
+++ b/requirements/requirements-dev.txt
@@ -1,5 +1,6 @@
-flake8>=3.8.3
-mypy>=0.782
-black>=19.10b0
-pytest>=6.0.1
-pymatrix>=3.0.1
+black==19.10b0
+flake8==3.8.3
+mypy==0.782
+pymatrix==3.0.1
+pytest-cov==2.10.1
+pytest==6.0.1