## Changes

* Addresses #3 
* Cache the parsed entity and constraint definitions at file read
* Lock the requirements
* Add code coverage
* Bump version

## Benchmarking
For the RSX rear suspension case. The motion is defined as `range(-50, 110, 5)`. Single threaded.

Before:
```
%timeit result = solver.solve_motion(MOTION)
1.02 s ± 11.6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After:
```
%timeit result = solver.solve_motion(MOTION)
755 ms ± 2.75 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

* 25% ms per loop reduction
* 76% ms per loop variance reduction